refactor: refactor mapper to util class
